Rename form_designer.py to formdesigner.py
So module name does not clash with view function name.
