+++ corehq/apps/app_manager/views/__init__.py
@@ -131,7 +131,7 @@ from corehq.apps.app_manager.views.translations import (
     upload_bulk_app_translations,
     upload_bulk_ui_translations,
 )
-from corehq.apps.app_manager.views.form_designer import (
+from corehq.apps.app_manager.views.formdesigner import (
     form_designer,
     get_data_schema,
     user_registration_source,
