+++ corehq/apps/app_manager/views/formdesigner.py
